+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [BREAKING] Improved store API errors (return "not found" instead of "internal error" status if requested account(s) not found) (#518).
 - [BREAKING] Migrated to v0.11 version of Miden VM (#528).
 - Reduce cloning in the store's `apply_block` (#532).
+- [BREAKING] Changed faucet storage type in the genesis to public. Using faucet from the genesis for faucet web app. Added support for faucet restarting without blockchain restarting (#517).
 
 ## 0.5.1 (2024-09-12)